demo/requirements.txt
Django==1.11.7
django-ajax-selects==1.6.1
